Migrate systemAudioActivated to HdmiControlService
Note that for all the devices that connected to the same HDMI system,
their System Audio Mode should be the same. So for devices that share
the same physical address, they should be able to share this status in
HdmiControlService. It's also easier for framework to set and get the
status.

Test: manul
Bug: 123369653
Change-Id: Id42995fb2c4687e71d98f431c2cc6a9ff9b53737

Former-commit-id: 489454ff187f732da86a8507fa7842e5b06ea822