Add notification for Test Harness Mode
As part of Security Review's requirements, I need to add a notification to
Test Harness Mode. It is visible iff Test Harness Mode is active (unless
the user blocks notifications from the system).

Currently, tapping on it does nothing; there's no Intent available in
the system to direct the user to perform a factory reset.

Bug: 80137798
Test: make && adb shell cmd testharness enable; verify that the dialog
exists
Change-Id: I9b391bce57ec1ef5392bd78bfdcf62a3a7e00e55

Former-commit-id: b5abc2f4510aa1772c1debed2f6f7731049c4593