Fix ImageReader#newInstace with usage
Need to actually respect the given usage or the API doesn't
work. Shift the default to be done Java-side where it's clear
if usage was user-provided or not. Remove the incorrect
validations and assertions around usages.

Add missing HardwareBuffer#isSupported to let users
check if a usage & format combination is supported.

Change-Id: I70bfe6e6c41e78425a38af6b17d421f0e80e818b
Fixes: 123423319
Test: HardwareBuffer CTS tests

Former-commit-id: 4d312b212c0c7fa7978c788e2a8f5b6b51621c9c