Take CLONE_TO_MANAGED_PROFILE into account in InputMethodSettings
This is a preparation to clone SHOW_IME_WITH_HARD_KEYBOARD to profile
users.

With this CL, InputMethodSettings takes CLONE_TO_MANAGED_PROFILE into
account when writing secure settings.  The point is that when a secure
settings key is in Settings.Secure.CLONE_TO_MANAGED_PROFILE, the value
needs to be written into the profile parent's settings, not the
current user's one.  In this way, InputMethodSettings doesn't need to
be updated when a new entry is added to / removed from
CLONE_TO_MANAGED_PROFILE.

This CL does not change Settings.Secure.CLONE_TO_MANAGED_PROFILE hence
there should be no behavior change.

Bug: 123379418
Test: make -j checkbuild
Test: atest CtsInputMethodTestCases CtsInputMethodServiceHostTestCases
Change-Id: Ieefefb8630ddef3b247ebb865a604e5c72dfb49c

Former-commit-id: 15be5e6f1c19ff329a6382c5063aa58bc92fb36e