Simplify IMMS#startInputOrWindowGainedFocusInternal() part 1/5
This is a series of CLs to reduce the average indentation level of
InputMethodManagerService#startInputOrWindowGainedFocusInternal().

Currently the method in questions has a relatively high indentation
level.  As a preparation to reduce the indentation level, this CL
moves user ID resolution logic from
  IMMS#startInputOrWindowGainedFocusInternal()
to
  IMMS#startInputOrWindowGainedFocus().

This is a mechanical refactoring.  There should be no behavior change.

Bug: 34886274
Test: atest CtsInputMethodTestCases CtsInputMethodServiceHostTestCases
Change-Id: I8aa758519bdf65cdb9ec5964f6b8e7b5a857e269

Former-commit-id: 4391c2047553c5100e2090b0afd233c3339da945