Simplify IMMS#startInputOrWindowGainedFocusInternal() part 3/5
This is a series of CLs to reduce the average indentation level of
InputMethodManagerService#startInputOrWindowGainedFocusInternal().

With this CL, Binder.{clear,restore}CallingIdentity() are also handled
by the caller side, which reduces the indentation level of
startInputOrWindowGainedFocus() by one level.

This is a mechanical refactoring.  There should be no behavior change.

Bug: 34886274
Test: atest CtsInputMethodTestCases CtsInputMethodServiceHostTestCases
Change-Id: I807608d52bd5240c235b8270132c9cc622092f9a

Former-commit-id: 13dc12f833ecbe81e5eded4a0b88b0466a101b55