AML: Handle NoDisplayActivities correctly
There is a race when trying to determine if the launching activity is a trampoline activity or if
its a NoDisplay activity that will never draw on screen.

If the activity becomes invisible, its visible state in ActivityRecord may not change to invisible
before the activity finishes. This is a quick fix to also check if ActivityRecord is finishing to
determine if the activity is no longer visible.

Bug: 80380150
Test: atest CtsActivityManagerDeviceTestCases:ActivityMetricsLoggerTests
Change-Id: I55251f03db7390d6a3465eff851c4635680a2e07
(cherry picked from commit 57215c7b8e5c096ce4117a2bd4669f621b7ccf57 [formerly 6397dd1c922697b4cf09dd23124f67d0578328b4])

Former-commit-id: b1de42de9fefc0ca81ce89da4468ec0e77dcc3ec