Revert "Fix FDE"
This reverts commit d610092230798bf4c36f28e56aefffd54a646c83 [formerly 54271901cd9ea10ff4a9ef71823db3a41e52ef2c].

Wifi service is assumed to be always up by other parts of the Android
system. So, not starting wifi service will cause NPE crashes all over
the place.

Wifi service already goes into idle mode to handle this FDE transition,
so there is no need for it to be turned off.

Bug: 78663946
Test: Manually verified non-FDE to FDE transition on Taimen (Using
ag/6180881).

Change-Id: I2d4d38e978a671f4dcd671524a1b38f557510bf5

Former-commit-id: c0bd945890c5d9369d25a238c02cb9b2c105eb3e