Fixed ContentCaptureManager.getServiceComponentName()
Here's the current code:

mHandler.sendMessage(obtainMessage(ContentCaptureManager::handleGetComponentName, this, resultReceiver));
return resultReceiver.getParcelableResult();

That code posts a message to a handler, than blocks waiting for the result. It used to work fine,
but now it deadlocks because the handler is also running in the UI thread.

The solution is to not use the handler anymore, at least for these non-session related methods.

Test: atest BlankActivityTest#testGetServiceComponentName_onUi
Fixes: 123521559

Change-Id: I774dc6338f685d185447cd61da31bcc63b085996

Former-commit-id: effa5f438369810a3381cec612ab02deeda92f4c