Move MediaMetadata class back to framework
Due to the @UnsupportedAppUsage on
MediaMetadata.getKeyFromMetadataEditorKey, we'd better not to move the
class to APEX. This CL moves the MediaMetadata back to framework.

Also, in order to avoid hidden framework API usages in mainline module,
this CL moves the usage of hidden constructor of MediaMetadata.Builder()
from MediaSessionEngine to MediaSession.

Bug: 123488217
Test: make update-api -j; make -j;
      vendor/google/tools/flashall;
      atest CtsMediaTestCases:android.media.cts.MediaSessionTest;
      atest CtsMediaTestCases:android.media.cts.MediaControllerTest;
Change-Id: I21b4f0f59214478071bb0eafa93bb9538e44bf4e

Former-commit-id: a17bf2d0f280024e9ba9761458a6cc301c31eb6f