Make the smart-replies 'enabled' key control both actions and replies.
The key named 'enabled' in the flag SMART_REPLIES_IN_NOTIFICATIONS_FLAGS
currently only controls whether app-generated smart replies will be
shown in the system UI.
With this change that key controls both replies and actions from both
the system (the assistant) and the app. So if the flag is turned off
users won't see any smart suggestions at all.

Bug: 122506860
Test: atest SystemUITests
Test: 'adb shell settings put global
    smart_replies_in_notifications_flags enabled=X' for different
    values of X and ensure smart suggestions are turned on/off
    accordingly.
Change-Id: Id6b1072f11bf1ec4e0b68ef2cefb7e28bb65746e

Former-commit-id: 1ee9c9d0b5f15f52446341c30c82ac00579dec94