Track and dump notification lifetime extension
Modifies NotificationEntryManager to track which notifs are have
their lifetime extended (using a map).

Slightly modifies the call behavior of setShouldManageLifetime(entry,
false). Before, we would call this on all lifetime extenders, even
those that were not actively extending the lifetime of the notification.
No existing extender was depending on this behavior (and it was
confusing), so it's been removed.

Bug: 123547965
Test: atest
Change-Id: I76be09d645ca614e2cbbc145fc50c3fcbbd120fc

Former-commit-id: a5dad556ba86db921961272dc8f53f3e255b0b99