Enable offload queue based on flag value
This flag was added in ag/6102001, and defaults to false.

This experiment flag is backed by a persistent system property:
"persist.device_config.activity_manager_native_boot.offload_queue_enabled",

It defaults to false, which will perform as if the offload queue was
never created, since no broadcasts will be placed on that queue.

For simplicity in implementation, we create the offload queue anyway,
but nothing will be placed on it, and every operation on that queue will
be no-op.

When the flag is true, the offload queue is used for long broadcasts,
which for now is BOOT_COMPLETED.

Bug: 120794810
Test: dumpsys activity broadcasts # does not have offload queue
Test: device_config put activity_manager_native_boot
offload_queue_enabled true, then reboot and check dumpsys
Change-Id: Iccd6c18b688a2c04eb5ec61d4a020df308dd14cd

Former-commit-id: 0193e955fef43ef07796280da27c88b7cd0fdd3a