Add blastula pool system properties
This patch adds the following properties to DeviceConfig:
* BLASTULA_POOL_ENABLED
* BLASTULA_POOL_SIZE_MAX
* BLASTULA_POOL_SIZE_MIN
* BLASTULA_POOL_REFILL_THERSHOLD

The BLASTULA_POOL_ENABLED property is checked by ZygoteProcess but not
currently used due to an existing bug (b/123409530).  This will be
enabled by go/ag/6162164.  Until then the code path is tested via log
inspection.

The remaining properties are checked by ZygoteServer.  Values are
clamped to ensure that they cannot be set to values that are harmful to
the device.

Since device_config is not available in the Zygote the system properties
interface is used instead.

Bug: 123524494
Bug: 68253328

Test: adb shell device_config put runtime_native blastula_pool_enabled true
Test: manually verify the property change is observed

Change-Id: I2b675afd9fbc1cbb0e8bc1c491cfdbbb612d0d3b

Former-commit-id: 0bccbf7984e482b66dd4f4f6561b25c6a65af37f