Answer "on" to TV querying system audio mode status when the audio system is still initiating it.
Note that System Audio Mode initiation from AVR side could take a while.
During the process, the internal sam status is false.

When TV queries the sam status, current framework responds with the status
at that moment. Which means it returns false when sam init is not done.

But since AVR is turning the feature on and it knows
that TV supports sam when receiving TV's query. AVR can answer with a
yes in this situation.

Test: manual
Bug: 123287727
Change-Id: I544d2a9af12b71c66cbc24327792993f621b2520

Former-commit-id: 27ae403dc07a8c8f409a2e69b7b0a760eedb0ea9