Track Multiple Event Instances in EventReporter
To avoid spamming users or the debug app with debug
events, the DebugEventReporter will now track each
event it receives and only send a single intent for
each event signature per boot. In the future, some
other method of persistence (such as per-build) might
be preferable, but this should mitigate any looping
events.

In addition, add dump() to the event reporter so that
even if there is no debug app installed, or in case
an event happens multiple times, it can be observed
through a bugreport.

Bug: 120941729
Test: dump and observe that the events are printed.
Change-Id: Iaf04a308a29bc074acfaa16b9e70947761759181

Former-commit-id: 821857f776da92438df05ae7794db264b580b30d