Do not generate smart suggestions if the last message is from local user
Also, instead of using Person.equals, introduce arePersonEqual. It is
because Person.equals is doing == check on the icon, which may return
false even the icons are actually the same.

And fixed a bug in handling sender == null in Message.
According to the javadoc of addMessage, person == null means the message
is from the local user.

Test: atest SmartActionsHelperTest

BUG: 123996228
Change-Id: I0ab24d53796d4b124c23cacf3df25cd8705d4fb3

Former-commit-id: 7ab8308992aae80736bd9114113401ca477dab98