Improve API documents of some IME query APIs
This is a follow up CL to my previous CL [1], which enabled following
IME query APIs to return the result based on the caller's user ID, not
based on the current IME user, even when they get called from
background users.

 * InputMethodManager#getInputMethodList()
 * InputMethodManager#getEnabledInputMethodList()
 * InputMethodManager#getEnabledInputMethodSubtypeList()

This CL just clarifies it in the API documents.

This is purely a document fix.  There should be no behavior change.

 [1]: I192a0f5a1375170d17a4c08af94f23966dbaea8b
      6a209b6f7870c123eda2ff2ddd2e30b20ae4641a [formerly 7f8ee4b9ddd31ad36a12c5278b27990dc76011cc]

Bug: 13002424
Bug: 122164939
Test: make -j checkbuild
Change-Id: I55f51e211a8a47c01469e2d93b04c5c864f0b9ea

Former-commit-id: 629271aa85327dca1bf800a55f48f606dbf504ba