Removing hidden api usage from AutoCompleteTextView
doBeforeTextChanged() and doAfterTextChanged() were being called to
trigger a refresh of the autocomplete suggestions when the refresh
was not automatically triggered by text input.

An androidx usage example is in SearchView, where we need to trigger
the autocomplete results manually.
https://cs.corp.google.com/aosp-kitkat/frameworks/support/v7/appcompat/src/android/support/v7/widget/SearchView.java?rcl=197708fc4fc0a6a5ede2da2479613382b561f028&l=1199
https://cs.corp.google.com/aosp-kitkat/frameworks/support/v7/appcompat/src/android/support/v7/widget/SearchView.java?rcl=197708fc4fc0a6a5ede2da2479613382b561f028&l=895

I feel this is a valid use case and added a new public api refreshAutoCompleteResults()

Bug: 123768711
Bug: 123768433

Test: Added new CTS tests for newly added api

Change-Id: Icfa6287889b4f63c00aa2ad6450e47942a9adda5

Former-commit-id: ccec37e1d13ae6b5223f9a29696d444ae6f3e1f2