Restrict access to View's hierarchy listener.
The private field should not be used directly. There is a setter for it
called setOnHierarchyChangeListener that allows to provide listener
implementation. The callback should not be called outside of ViewGroup
as ViewGroup also calls onViewAdded / onViewRemoved. In case the
reference to the callback is really needed, the developer can override
the setter.

Bug: b/123768704
Test: N/A
Change-Id: Idfc9c656d496b4e3e2c134a1d5e2c95f15099f51

Former-commit-id: cfb30af1182efed17a61a1c6848d5e1bc9060f1d