Restrict FLAG_SUPPORT_STATIC_TRANSFORMATIONS usage
There is a proper setter for it already called
setStaticTransformationsEnabled. There is not a strong motivation for a
getter as this is only used to instruct the class to invoke
getChildStaticTransformation when a child is drawn. And any class
setting this flag to true is supposed to override that getter. Also I
was not able to find any reasonable public usages of the flag.

Bug: b/123769647
Test: N/A
Change-Id: I14ea7d560a5318cc372ac84dc5a1daad626e4f7e

Former-commit-id: d590b1a719f0dbb992812dc064fff332e3de276f