Forward StatusBar.onBackPressed to Bubbles when expanded
StatusBarWindowView is the root of the StatusBar window
and handles all key events at dispatchKeyEvent

All keys dispatched to the StatusBar window flow through
StatusBarWindowView#dispatchKeyEvent. A BACK key event
is forwarded to StatusBar.onBackPressed()

If StatusBar#canPanelBeCollapsed() is false, the action is
passed to BubbleController which forwards to the expanded
bubble by calling to ActivityView#performBackPress().

Bug: 123631742
Test: manual, go/bubbles-app
Change-Id: I0ccafc02f876c973ebc267055162621109b2dcc0

Former-commit-id: 041d72697a31c0f143804607a186554581a1f53a