Make RCS Message Store APIs ready to unhide
This change makes necessary changes to unhide the APIs in a future
commit. All the classes that end with Parameters needed to be converted
to Params, which resulted in a rather large change. I also had to reintroduce
RcsOutgoingMessageCreationParameters as it was required to make
Parcelables final, which means I had to make this new class parcelable
and RcsMessageCreationParameters non-parcelable and non-final

Test: Existing tests pass

Bug:109759350
Change-Id: I2fe9382d28a9abf02ad22ec7ccdeaedc55ae4337

Former-commit-id: 2f932d72baf5cb831c973c838505c34ecb84b64c