Restrict reflection of default DisplayInfo constructor to P and lower.
There doesn't appear to be a valid use to create this object yourself;
you can retrieve DisplayInfos of the attached displays using various
other APIs, such as through WindowManager.  The copy constructor is
also available.  This method is only used by one app.

One possible usage is to get the display cutout information; this is
something that is now available as an official API in Q.

Test: N/A
Change-Id: I632b337e5dfdcddd494b03fd1403f73d2c46a475
Fixes: 123769467

Former-commit-id: 47ed9eec7670d4e045dda71695a26bfc0a887f99