Update GNSS listener maps when clients die abnormally.
In the LocationManagerService class, the GNSS measurements
and navigation message listener maps are not updated when
clients die abnormally without calling unregister callback.
This causes the maps to grow and contain invalid entries.

Also, the GNSS status listeners are not managed the same way
as the measurements/navigaton message listeners (b/120449926)
Fixes are somewhat related hence better to do in the same CL.

Fixes: 119584292
Fixes: 120481270
Fixes: 120449926
Test: Verified with GnssLogger using the repro steps mentioned
      in b/119584292 and b/120481270.

Change-Id: Ibf0283d7e8b2af09e19892d2e22d7f3650420033

Former-commit-id: 98d49b7ff6f5e895890c646bec050e8e1e2c7904