Open API for onCreateView and createView with Context
Fixes: 123768619

LayoutInflater's mConstructorArgs was being accessed for two
reasons: reading and writing the context. When onCreateView()
was called, the inflation context was not being passed, so
if the developer wanted to get it, they had to read it from
mConstructorArgs. When the developer wanted to create a view,
with createView(), the developer could not adjust the view's
constructor context.

This CL creates versions of those two classes that take
Context as a parameter to remove the need for accessing
mConstructorArgs.

Test: ran LayoutInflaterTest
Test: Ia40341c24998be13205a72386d217d038a3bcf93
Change-Id: Ic2e990567512dd051f3d4f6c842398c71d6817b5

Former-commit-id: f38b2c62ab78a50f03b68433f5aec93ebc179430