[GPU Service] Move blacklists process to GPU service.
Instead of re-process blacklists proto everytime, we move the processing to GPU
Service, and add SettingsObserver to observe the change of
Settings.Global.GAME_DRIVER_BLACKLISTS such that we only re-process the
blacklists when needed.

As a result, we will have GAME_DRIVER_BLACKLISTS which is used to accept a list
of blacklist from the server, one for each game driver version that needs to
have blacklist, and GAME_DRIVER_BLACKLIST will only contain a list of
blacklisted application package names for the current version of game driver on
the device, separated by comma.

BUG: 120869311
Test: Build, flash and boot. Use adb shell settings put command to verify.
Change-Id: Ic49800cd0b5b82ddbbdf9907d603243415d5bd3b

Former-commit-id: 50350c15b8f439a42cf9d359d3c98a91fdd7966e