Fix flaky WindowStateTests
The test testVisibleWithInsetsProvider was flaky because the provider
cannot control the visibility if the USE_NEW_INSETS_PROPERTY in system
property is not NEW_INSETS_MODE_FULL. In order to test the provider
visibility controlling functionality, we need to make the top bar insets
controllable.

To do so, we need to update sNewInsetsMode in ViewRootImpl for the
test.

Test: atest WindowStateTests
Bug: 74078662
Bug: 69229402
Change-Id: If6aba86e3c5a55dd0976ddda3a93f003ebaf5712

Former-commit-id: 5e696f1c642197331d75edc9b8dc3a715db4b077