Disallow access to IGraphicsStats$Stub starting with Q
Reasoning: this is deeply internal apis and it is impossible
to use them without calling tons of other private apis.
Also, there is no known valid use cases, when it would be useful to do.
If you need graphics stats for debug purposes, please use
"adb shell dumpsys gfxinfo <package>" and/or systrace


Change-Id: If04c7802c9e663f62b85c23d81fa11514137319b
Fixes: 123769511
Test: N/A

Former-commit-id: 126677222c4d6429f0584fdfe3002d2ec4e202c5