Restrict reflection access to NumberPickers's mSelectionWheelPaint
Applications reflecting into mSelectionWheelPaint do it because there is
no API for correctly updating the text size and color at runtime. Since
the NumberPicker is composed of an EditText and an additional paint to
draw the scroll wheel numbers, developers found themselves reflecting
on the paint since updating the EditText child is not enough.

This CL restrict access to mSelectionWheelPaint but add an alternative
API for updating the NumberPicker's text size and color at runtime.

Bug: 123769712
Test: NumberPickerTest
Change-Id: I9f5a0275916a6852a6378e5ffae5a9a171b9cd51

Former-commit-id: 11a57f22b4b81c341ba7f5f375f9dfba1bbabacf