Mark FrameMetricsObserver::notify() as pure virtual
Bug: http://b/116873221

If not, Clang assumes that the TU which defines this function will
contain the vtable for the class as well.  Since there's no out-of-line
definition of this funciton, no TU ends up with the vtable.

This causes a problem with coverage builds, which are built with -O0,
where calls don't get inlined, thereby requiring a definition of the
vtable.  For non -O0 builds, the vtable is not required since the
virtual calls get inlined/optimized-out.

Test: Build with and without -O0.
Change-Id: I60a5cefcd1c327f1e00785fedbb2163c682b33d1

Former-commit-id: 459afe96d13876c1fbeacb1f0d4bd0eddb7342e0