Ensure setActiveGroup() does the right thing after a HAL crash.
Before this change, it wasn't deterministic to reset the HIDL layer
after a simulated HAL crash.  It looks like it required a second client
to connect before onError() ultimately corrects mCurrentUserId.
This code forces mCurrentUserId to an unknown state, which ensures it gets
reset when we re-connect to the HAL.

Fixes b/124075073

Test: HAL crash resets mCurrentUserId.
Change-Id: Icca2ad2463dedaf0834e61c744a2ddcac15f6356

Former-commit-id: 7b78b22ac1d8f2c7685b149a3b6767ddeea6f3f5