Write ADB keys to adb_keys as well for backwards compat
The original design of revoking adb keys after a period of
inactivity did not properly verify that the system held the
corresponding private key. There will need to be additional
coordination between adbd and the framework after a system
passes the SIGNATURE challenge to determine if the key should
be allowed, but this change will restore the original behavior
for beta 1.

Bug: 124076524
Test: Manually verified key is written to adb_keys and subsequent
      connections went through the SIGNATURE challenge

Change-Id: Ib735f8a4d83518d4f6b5e939dcd1a2f5ae6f4837

Former-commit-id: c5da14c2252d363788f65f623949b8ae53e4bb27