Updates individual bubble dismissal logic and animation.
This makes it so that bubbles can be dragged out, and the rest of the stack will animate to close the gap. If the bubble is dismissed (either by a hard fling or by placing it in the dismiss area), it'll animate away. Otherwise, it'll animate back into the row of bubbles.

This CL also simplifies the touch handler code. Along with eliminating some redundant paths in onTouch, this also deletes the FloatingView interface. Now that the stack and individual bubbles act so differently, it makes less sense to have a single interface for moving them.

Test: atest SystemUITests
Bug: 123541855
Bug: 123022862
Bug: 124058992
Change-Id: I57e3df19a191a6da0b5ed5d40c350bebb29f74ff

Former-commit-id: 442b62779bf15ec9a7a5ca2942e04bc317c172bd