Allow reflection access to ProgressBar#mOnlyIndeterminate.
I think this one is fine to punt.  The field applies to the default
progress bar, which can only be indeterminate.  However, if someone
can't customize their themes and styles, they may end up in a
situation where they need this to be different and there is no public
API for it.  Making a public API is messy because it is really hard
to reason about and the behavior would be a no-op much of the time.
It's better to avoid adding confusion for the majority of developers.

Test: N/A
Bug: 124049927
Change-Id: Ie63429831ec803d2d3c50464176c4e6f8a30ec33

Former-commit-id: 08b675370ad478e71103324f8133b416490810fd