AccessibilityRequestPreparer shouldn't crash if the view is released before the preparer is added or removed.
AccessibilityRequestPreparer holds onto a weak reference of the view.
When it is added or removed, AccessibilityManager uses the views accessibilityId to maintain the lists of requestPreparers.
But, it is completely possible, that a view is released before requestPreparer is removed, or even added, and that will cause bad things.
Instead, store the id on the preparer also.

Change-Id: I5d489c061cd8039d066a81f4a927c1d8185d4f06
Fix: 123047944
Test: CtsAccessibilityServiceTestCases

Former-commit-id: 1c88bcee5c834f2dfd40b7a862c0ec80074e7e47