Modem power monitor
This is an AOSP carry over for the frameworks base CL. This change is needed here to allow telephony changes to be added to AOSP with no issues.

Adding monitored power rail as a new cellular battery stats component. This will be used to better determine actual power usage of the modem rather than relying on battery reduction rates.

If the device does not support rail power monitoring then this value
will always remain at 0.

Test: adb shell dumpsys activity service TelephonyDebugService --metrics
Results Examples:
	Energy consumed by modem (mAh): 2.41

Change-Id: If85af2f2e1b255661e189c5375ecc5825be84c58

Former-commit-id: 4d3a9f2b3c9873956d829faca9f797d9a69eb30b