Revert "Revert "[GPU Service] Move blacklists process to GPU service.""
This reverts commit 0833a50e0dba7ee34d4a3eec300027ab41fb2c94 [formerly 67773e8b878c9c789c56f96f7efd23ef9e710249]. Below is the
original commit message:

Instead of re-processing blacklists proto everytime, we move the processing to
GPU Service, and add SettingsObserver to observe the change of
Settings.Global.GAME_DRIVER_BLACKLISTS such that we only re-process the
blacklists when needed.

As a result, we will have GAME_DRIVER_BLACKLISTS which is used to accept a list
of blacklist from the server, one for each game driver version that needs to
have blacklist, and GAME_DRIVER_BLACKLIST will only contain a list of
blacklisted application package names for the current version of game driver on
the device, separated by comma.

BUG: 120869311
Test: Build, flash and boot. Use adb shell settings put command to verify.
Change-Id: I3a82d1cc020640f7b595def67b280dbc2d3a3155

Former-commit-id: 248542053b1cac7ef141a8292a521862185498a8