Set MediaPlayer sources in more direct fashion.
As part of the isolated storage work in Q, some filesystem paths
don't actually exist.  It's safer to simply try opening the file
and let FileInputStream throw a FileNotFoundException.

This change results in more descriptive errors to developers,
without changing the behavior of throwing an IOException if the
file doesn't exist.

Bug: 124208701
Test: manual
Change-Id: I4710a55431d8dc6cb96561f551b34892d8c15ba5

Former-commit-id: a0a955f814307c65c872ad6d02e008e6ad2b5a6e