Allow framework-atb-backward-compatibility to be added to bootclasspath
Fixes a build failure that occurs when it is executed with
REMOVE_ATB_FROM_BCP=true.

Setting REMOVE_ATB_FROM_BCP=true adds the
framework-atb-backward-compatibility target to the PRODUCT_BOOT_JARS
path in place of android.test.base. That causes a problem because the
build expects that everything on the PRODUCT_BOOT_JARS is installable.

Bug: 73711752
Test: make droid and flash both with and without REMOVE_ATB_FROM_BCP=true
Change-Id: I6595ff8dba42e6b85abbf6a6cad4d8970b5a91a0

Former-commit-id: fab8754054e3a87600bb490d1d0cc81b1eb092c9