Added visual affordance/handle for navigation bar
Added the handle that would sit at the bottom where the navigation bar
would be and off by default. Use prototype to make it visible. The bar
will also use color adaptation. Needed to fix null states when the
inflator would swap the normal home/back button layout with the handle.
Added the new 16dp overlay as the visual inset.

Test: manual
Bug: 112934365
Change-Id: I88f3443ad36ceefa8f490d17b4daefbeaab5c60a

Former-commit-id: b55c02c9dafacb2e494cb7b438bf05776c83033e