Make postively z-ordered sub window above preserved surface
This original logic moves the preserved surface of an activity window
to the same layer of the first child window while tapping on the
divider in split screen mode. The preserved surface covered the child
before being destroyed, so that it looked like the child window was
blinking.

The preserved surface is at layer 1, so this CL makes all the postively
z-ordered sub window be placed above that layer.

Bug: 124732440
Test: atest WindowManagerSmokeTest ZOrderingTests
Change-Id: Ib2627e3007ac88bf4b2936f412efb4803ed86007

Former-commit-id: 9c8ee26b7441b9989c23069eaac43e77b66a4913