Diff resource dirs when checking LoadedApk packageInfo cache in ActivityThread
Need to check the overlayDirs with the current appInfo passed into
getPackageInfo, otherwise it's possible to return an old, outdated
LoadedApk which hasn't loaded the correct resource dirs into its
AssetManager.

There is a separate issue with AssetManager2.isUpToDate not being
implemented correctly, but that will be moved to a different
changelist.

Bug: 124363683

Test: manually tested Wellbeing with repro steps in linked bug

Change-Id: Iace6e8aab322bdb7b5b93398a4f5e2c69fccc68e

Former-commit-id: 83708c825868b4c26c6c599b5614205ea5118d3c