Move PublicFormat definition to libui
PublicFormat is defined in android_view_Surface.
However this and related functions are also used in libmediandk
which is both NDK and LL-NDK. As a LL-NDK it should run without
libandroid_runtime.

By moving PublicFormat definition and its utility functions
into libui, LL-NDK can use it without cost of libandroid_runtime.

This is a part cutting dependency from libmediandk to
libandroid_runtime.

Bug: 124268753
Test: m -j
Change-Id: I49eea4325cde244a7168c6954027ee5c9a8459da

Former-commit-id: 63f5f927fcc30882aa53c2975e27f35c49af0639