Added pool management code around zygote state changes.
The hidden api access log sample rate and api blacklist exemption
commands change the state of the zygote and processes spawned from the
zygote.  This means that blastulas spawned before the changes won't have
the proper state.  To fix this we now empty the process pool, apply the
changes, and then refill the pool.

Test: Build and boot
Test: Treehugger
Change-Id: I66b01930a160c1305fbeb8b6243deff7917aa598

Former-commit-id: 01060e66f6a91859108df0482a31b5b6c53598e5