Extend PackageWatchdog with explicit health checks
Allow PackageWatchdog to monitor packages with explicit health checks
enabled. In this case, at the end of a monitoring duration if a
passed-health-check callback is not triggered, the package would be
regarded as failed (in a later cl) and the observer is notified.
If monitoring without explicit health checks, the behavior is the same
as before, packages expire silently.

TODO: Implement the package failure trigger on expiry with failed
explict checks and enable added tests

Removed username from TODO comments

Bug: 120598832
Test: atest PackageWatchdogTest
Change-Id: Ib4322c327bcb00ca9a3fbdc83579e7b5f2fd633b

Former-commit-id: 9284e74585971c690951b398a3c8c2d3322df5af