Use inner frame of letterbox as touchable region for letterboxed app
Since commit 192fe76a, the letterbox is touchable. In order for the
touch event to be able to slip into the activity, the touchable region
of activity must not cover the letterbox.

Also toggle the token of window handle with surface show/hide to reduce
the inactive window records in InputDispatcher.

Bug: 110257889
Test: manual - Enable display cutout, open DocumentUI in landscape and
      drag from the black region in the cutout side to DocumentUI, the
      list view should scroll or the drawer slides out.

Change-Id: I4a0bffebfe7a3efcac842c0abbae60ea2a3487bd

Former-commit-id: 69de5fbb71c22ee5bcc7773c1da33452b5843edc