Revert "Revert "Remove legacy-test and legacy-android-test targets""
Fixes up the android.test.legacy target to stop using sources directly
now that the issue with android.test.runner-minus-junit not being able
to be built against the SDK has been resolved.

Also, fixes a couple of comments.

This reverts commit 1158b1adaa9e0d0144a4c142c0fbe1238110c5a1 [formerly 7d7af10dc9057a49c617c7c97c174e8b4ec5008c].

Reason for revert: When this was originally applied it broke the
pi-dev-plus-aosp branch as it contained references to these targets.
Those references have now been removed.

Bug: 63127064
Change-Id: I619e92faf4cc38f8fd4f053bde832e9a76a2a2bf
Test: make checkbuild

Former-commit-id: 79d7ba2773e5e6eed4c8176ebaf8c4199839b343