Use new libcore filegroup variable
This change switches frameworks/base/ targets to a generic "interface"
libcore filegroup variable, core_public_api_files, rather than having
it know about Core Library internal filegroups. It is not perfect
because the command line tools need to know to exclude certain packages,
but it is a step in the right direction.

Bug: 126332291
Test: build only
Change-Id: I72667f30dc5685fdc24aca3f8ee82b20c7548449

Former-commit-id: 3b1b70376cd3033ab587f0fc469b7a4503cf8bd9