Added a filegroup for statsd config proto definition.
This is used for a testing framework that involves serializing text
proto statsd configs at build time. The filegroup allows aprotoc to
access the definitions elsewhere in the source tree.

Submitting this in internal first as the AOSP version of statsd old and
causes merge conflicts when submitted there directly.

Bug: 120569785
Test: Build is not affected.
Change-Id: I0e37a9b9a1f22a5dcebbf07ccc5cad98cf76024c

Former-commit-id: fe16c8f2f90c4e80125ad516e82d2c1c53848ff5