Check if idmap2 is accessible before executing it.
When wrap.sh is used in a debuggable app, AssetManager tries to execute
idmap2 in the app's security context, not in zygote context. And idmap2
isn't allowed to execute in the app's context, thus making an abort.

The abort makes Android studio can't debug the app. To avoid the abort,
check if idmap2 is accessible before executing it.

Bug: 124109020
Test: test manually that android studio can debug the app with this
Test: change.

Change-Id: I82d772358b2266433f82319896a4ca662cc3c273

Former-commit-id: a60dc9e51dd02b5394523934f1a40434d67afab2