Display browsable layout for http/s scheme intents
Displays Settings button instead of always button for all intents that
have http or https scheme, as opposed to only intents that have
BROWSABLE as category.
This fixes an inconsistent and confusing behaviour that users were
experiencing, as sometimes they saw the Always button and sometime the
Settings button.

Fixes: 126499502
Test: open a mail.google.com from Facebook Messenger - used to display
Always button and now displays Settings button
Test: use shell command to send https intent without BROWSABLE category
- used to display Always button and now displays Settings button

Change-Id: I1084fd716ba8e584c2b17d8fe1500dbc641cab46

Former-commit-id: 0c7c5d59f8002500330b1571f9e5221d4316492f