Call into role controller asynchronously when holding mPackages.
For other cases calling synchronously is better, because callers might
expect the next get call to return the right browser. However in the
case of installing a browser app, we are holding the mPackages lock
and should not be calling into higher level components synchronously,
so use a new async call for this. This fixes the system-hang until
timeout when installing a new browser app.

Bug: 124452117
Bug: 123775970
Test: manual
Change-Id: Ib820e65c79c2315f41ff0e31268631c973af4511

Former-commit-id: fafe525798e6fce1934e3a8d98cfddb141050ca3