Make mIdsToViews thread safe.
View calls like "attachedToWindow" and "detachedToWindow" can be called
off of the main thread, on "android.anim" this happens in the system
process for application splash screens.
Let's wrap it with a synchronized block.

Fix: 124110476
Test: The crash is a race condition, so the best I can do is make sure
it generally doesn't crash anything by trying out both normal and
virtual hierarchies in Talkback.

Change-Id: I19f972a97b5ac80917f8dd77ac79bf7e3d5e826c

Former-commit-id: 1aa9da4250326da33e9666aa685ef747ff6ee40a