Small issues with FLAG_PERMISSION_USER_SENSITIVE_*
Add them to the "all flags" mask. We cannot change the value of a
constant in the API, hence need to deprecate it. This change does
not create a new API with the new value to not have the same problem
again.

Register the flag type in the IntDef and toString methods.

Test: Set permission flags, rebooted and verified that flags get
      restored.
Bug: 124317989
Change-Id: Ia68053a937be7dc47af6e313e66e87f3228a2d14

Former-commit-id: 7659769d3cc0009c84d6ae63c1af489b89f9766b