Revert "Update VPN capabilities when its underlying network set is null."
This reverts commit e345c9982e00d6b751f928151b2cb6ddd80ea98c [formerly ad8805a6c217338b12511f220cce22211a19080a].

Bug: 126245192

Reason for revert: This change can lead to a deadlock that was fixed in http://ag/6580635. However, platform PMs think that fixing this is risky enough as this is not a recent problem and has been in the field for 3/4 of the year.

Note: The merged-in tag is used to avoid this change from getting merged into pi-dev-plus-aosp. This is to avoid merge conflicts since we mostly work in aosp/master which merges into pi-dev-plus-aosp.

Change-Id: I3814bcec87efb059f50f00617406501aaeac3b4d
Merged-In: Id0abc4d304bb096e92479a118168690ccce634ed

Former-commit-id: 6a050c7c50fa0838d7e29b8c6e244018044246db