Fixed flaky testTaskChangeCallbacks and promote three tests to presubmit
The testTaskChangeCallbacks was flaky because the callback onTaskRemoved
can be called after finishAndRemoveTask is called. However, the original
test will assert the callback was not called after calling the
finishAndRemoveTask and wait for the callback afterwards. To fix it,
only ensure the callback was not called before the finishAndRemoveTask
get called.

The other three tests are no longer flaky and promote them to
pre-submit.

Test: atest TaskStackChangedListenerTest
Bug: 123654474
Bug: 74078662
Change-Id: Ie064ae00a4b2493e5aa0f47e1418ce2d25b503de

Former-commit-id: c1f3e3787027c40e402b7e27aa3ac0eb3ef73f92