Allow transitioning through lifecycle when finished
When finish was requested for an activity, certain
lifecycle transitions were blocked to avoid delivering
invalid callbacks in the past. The checks and early
returns that were left from the old way of resolving
lifecycle cancelled delivering STARTED, RESUMED and
PAUSED state if an activity called finish().
This is no longer required with Lifecycler infrastructure
that takes care of ensuring validity of all transition
sequences.

Bug: 127106719
Bug: 126693030
Test: ActivityLifecycleTests#testFinishInOn*
Change-Id: I5ed006f1846e54589f33da0cab3b6f3bf9ca5a61

Former-commit-id: f2d598e1684a8b3ed793611f878fafc4f126e1ea