Removed duplicate Passpoint entries in saved networks page
The AccessPoint constructor for PasspointConfigurations was not
setting the AccessPoint key. This caused the Saved Networks Page to
not be able to correctly reference the entries for saved Passpoint
networks, causing entries to not disappear after clicking "forget",
or creating duplicate entries whenever the activity was resumed.
Fixing the key updating/getting for PasspointConfigurations solves
this issue.

Bug: 118705403
Test: atest AccessPointTest
Change-Id: I0e49192ef5efa72fd490e4a0ac987e1c0682e3f3

Former-commit-id: 6e711c80ffaf97eb68066394ee730855519e5f9b