Mainline logging to statsd: native static libs
This is a cherry pick to aosp.

Creates an annotation to specify if an atom needs to be logged from a
specific mainline module. Creates options in stats-log-api-gen so that
if the same module name is passed in, cpp/h files will be generated with
only those atoms that are specified. These files can be used to create a
static library per mainline module to log to statsd.

Test: builds
Test: atest stats-log-api-gen-test
Test: does not affect existing shared libstatslog
Test: manually created a static lib for netd and used it to log to
statsd. Used testdrive to validate the data was properly flowing.
Bug: 126134616
Change-Id: I78064f81fb4971eede8e97dacce5424e3eefd8bb
Merged-In: I78064f81fb4971eede8e97dacce5424e3eefd8bb

Former-commit-id: 1f431f31e037456d5de3d8f660e1d9e59c6508e1