Added correct TalkBack support for QSCarrier
Added new MobileIconGroup to TelephonyIcons for SIMs that are not set as
default for mobile data. This group behaves identically to DATA_DISABLES
except that is checked before it so we can assign it correctly.

This addition was made for properly providing content description of
carriers that are not set as default for data. Previously it said
"Mobile Data Off" which could be misleading.

Test: manual with 1 and 2 sims
Test: atest NetworkControllerDataTest NetworkControllerSignalTest
Bug: 123299825
Change-Id: Ia9e085b6296ff8df861e1e40bf8004b20b0d2c11

Former-commit-id: c7bc84b92d39cf93466307bfef33533d11ea268a