Should return the number of supported SIM cards
When multiple SIM is enabled, getSimCount() in TelephonyManager.java
just returns 2 and it is not consistent with getPhoneCount(). For
devices which support more than 2 SIMs, it will cause
ArrayIndexOutOfBoundsException in the getInstance() function of CatService.java
when thress SIMs are present. To fix this issue, we need to return
the phone count in getSimCount() to synchronize both APIs.

Bug: 126292874
Test: check if JE is fixed on a device which supports 3 SIMs
Change-Id: I227f7d05349c1d0cf4c8bfa46c4e3b5f9bd42be9

Former-commit-id: f0026eebfc2849c51232793707721338a2e06883