Revise BroadcastRadioService handling of V2 HALs
Previously, BroadcastRadioService checked for HALs (both V1 and V2)
exactly once, the first time listModules() was called. This approach
failed to handle HALs that registered after boot or that died and restarted.

This change addresses both issues by registering a listener that is
invoked when V2 HALs are registered. Additionally, the check for V1 HALs
now occurs at construction time.

Bug: 121006324
Bug: 126887436
Test: Verified on device BroadcastRadioService handles V1 HALs properly and picks up V2 HALs dynamically
Change-Id: Id9df7f53d189801f0a0dc894b2bb425d655edbd6

Former-commit-id: 5af72432bb58d3d935cab7677e03197eb1e930c5