Only show the toast of dismissing docked stack if it actually happens
When launching a non-resizable activity on secondary display, an overlay
activity will show "App may not work on a secondary display" on it.
Because the overlay activity is resizable, it is nothing to handle on
secondary display. But it falls to the condition of split-screen, and
even there is no docked stack, the toast of dismissing docked stack is
still shown.

Bug: 128674132
Test: atest ActivityStackSupervisorTests# \
      testHandleNonResizableTaskOnSecondaryDisplay

Change-Id: I1b8fe892ec4cace4f87112f132e0eefdae19024d

Former-commit-id: bed3db4d0b888240a770bb57fa270569041f3994