LongSamplingCounter - remove update()
It wasn't being used anywhere and was requiring an extra long. Since
LongSamplingCounter takes up (cumulatively, due the huge number of
them) a lot of memory, each byte counts for a lot. (It was added in
ag/3771500, but appears to have never been used, so we restore its
original algorithm.)

Test: atest LongSamplingCounterTest
Change-Id: I7349e927ae6fe8c941e408c4d0843b638bf070e9

Former-commit-id: 1371b9ac52d6dd1f955c6bc6f8f04ad2b8e0785e