docs: Fixing {{ in code comment (which makes devsite publish choke)
One of the code samples in a Javadoc comment contains a {{ ;
this gets interpreted as a Django directive by our docs publish tool
(go/devsite2cmd), which results in a syntax error. We verified
(http://cl/238522152 ) that escaping one of the {s resolves the
issue.

Tested by building the docs and staging to:

http://go/mivi-stage/reference/android/net/wifi/WifiNetworkSuggestion.Builder

Bug: 128536865
Test: make ds-docs
Change-Id: I70a3e0481ad195665c878c72d0ec6b1ccbbf958f
Exempt-From-Owner-Approval: Doc-only change

Former-commit-id: 68bec891b49665ca12f60a0bac670d756ef15b75