Include language information in intents from the TextClassifier.
This CL runs langid on classifyText() and sets the language information
of the classified text in Intent extras that are sent to the destination
app. If the text is too small for langId to detect the text reliably,
include surrounding text as necessary.

Includes flags to tune/disable how sorrounding text is used.
Also fixes failing tests.

Bug: 126453204
Test: atest android.view.textclassifier
Change-Id: I9ef2a33690dc3787f6544ceed0f41c527f4423e5

Former-commit-id: 72e17976f38403a05a62d792613dbb296dcc85fe