Reduce the window tap exclude region for child above it
For letting touches directly go to the embedded display, we've opened a
tap exclude region for the ActivityView. However, if there is a view on
top of the region, the view cannot be touched within the region.

In this CL, we reduce the tap exclude region if there is a can-receive-
pointer-event view on top of the region.

Bug: 128517544
Test: atest CtsActivityManagerDeviceTestCases:ActivityViewTest
Test: atest FrameworksCoreTests:ViewGroupTest
Test: Menual test with ActivityViewTest and Bubbles
Change-Id: I68e2a9fe9d0891801b533ab8d25074f64bef5a79

Former-commit-id: 2b210c234c32b1ea3d760f018274576d05b24ad6