Deprecate boundsInParent in favor of boundInScreen.
If the service has boundsInScreen the boundsinparent can be computed.
Also for its whole existence, boundsInParent is the bounds within the
ViewParent's coordinates. But the ViewParent is quite often not the
AccessibilityParent (if it isn't important for accessibility), so the
service can't really do anything with that information because the
service doesn't have access to the ViewParent, only the
AccessibilityParent.

Test: It builds.
Bug: 1084662
Change-Id: If88dcaac6dcb824d72d7f483c05cbb544314423c

Former-commit-id: 31d0cedcc40df60359e28799925b21aa2b48840e