WifiNetworkSuggestion: Change the polarity of priority
The current ordering of priority is hard to handle inside the wifi
stack because "-1" was unassigned priority and 0 is considered the
higheest priority level.
Reverse the polarity so it's much easier to compare network
suggestions with different priorities such that
priority1 > priority2 is always true. unassigned priority will always
be lesser than any assigned priority value.

Bug: 127644378
Test: Compiles
Change-Id: I74dacdb85838250c242257a80ebbe6046cef752f

Former-commit-id: fd2df9552233a78a9962a3e40465d15df1e0a821