Sysconfig takes precedence over update-time autoVerify
autoVerify is re-evaluated whenever an apk is updated, and in the normal
case a verification failure here results in the app being demoted out of
'always' status for handling web link Intents.  However, bundled app
policy for this is supposed to be set statically via sysconfig --
so now we make sure that sysconfig policy takes precedence in this case.

Fixes: 127838380
Test: manual (update bundled sysconfig-cited app and force failure)
Test: atest OsHostTests#testIntentFilterHostValidation
Change-Id: If80511ce47847bb21a48634b191acfacda606331

Former-commit-id: 237aab5aff2adb8ff092d00f42682b09d0e51c79