Fix TB bugs in QSCustomizer TileAdapter
* When saving the specs on exit, make sure that TileAdapter is left in a
good state. This meanse removing the placeholder if it exists and
setting it to ACTION_NONE
* Update mTileDividerIndex when adding and removing a place holder. It
doesn't need to be updated manually after a completed add, move or
remove operation as we call #updateDividerLocations in #move
* When adding a tile in the position before last, it gets added one
position before that. This happen because of the fix for off by one that
is not needed anymore.

Fixes: 129059015
Fixes: 129001337
Test: manual using TB. Adding, removing and moving tiles, in particular
to positions near the end.
Change-Id: I010d118b15c1cad3c0ab0544b515eb9c5b53b2ce

Former-commit-id: b6a127f9509e06c9c0baf04407475148b305a452