Clean up references to @ColorLong
Bug: 120904891
Test: make docs && look at the output

Remove instances of {@link ColorLong}. Although the link is supposed to
go to
    https://developer.android.com/reference/android/annotation/ColorLong.html
my web browser redirects me to
    https://developer.android.com/reference
Perhaps there's a bug somewhere in the generation (I am able to navigate
to
    https://developer.android.com/reference/android/support/annotation/ColorLong?hl=en
by searching on developer.android.com - notice the addition of
"support"), but I still think it would make more sense to send users to
the page for android.graphics.Color directly.

Switch over to {@code ColorLong} (which does not generate a link) and a
reference to look at the {@link Color} class.

Update parameter named "c" to one named "color".

Unify various Canvas#drawColor APIs to say "to draw onto the canvas"
(previously some said this, while others said "to draw with").

Change-Id: I78eff5c35e350f763aac0416f6ec1ccf12f89418

Former-commit-id: fda4f610605a3e9330f1a4a66d435902cdcaf206