Null AppWindowThumbnail SurfaceControl after releasing it.
Otherwise SurfaceAnimator#reset may try and use it and throw an unhandled
exception. It looks like this should be safe because we always null the
SurfaceAnimator after releasing it's surface. The only other point holding
on to the surface should be the animator, and the animator properly checks
for null.

Bug: 124877830
Test: AppWindowThumbnailTest
Change-Id: I3841b27cc9e24ed24951df27b2b35a22bca70130

Former-commit-id: ad30f958a62f17c7c39b6c02b8028198501a3807